Options Screen Settings Implementation Start
During this arc I've started the implementation of Option List. I am using ListView to visualize the Options List on the the screen. Also I am using the ListDataObject to create a multiple types of Options that would be available inside Option List. Each ListDataObject has it's own unique feature, that would be implemented in future. Next task is to implement first real type of Widget and correctly display it in Options List.