Display of an actual content inside Option List
During this arc I've implemented the possibility for us to display a desired content inside the List Entry Widget. Now we can see a "Display" Setting Name, as well as current value of a Rotator. Also I've setup the functionality of the Next and Previous Buttons. Now after we press either of buttons we will get next property from the Data Object ( if there any ) and display it in our Rotator. If there is no values in Data Object, out Rotator value will stay the same.
Next task is to implement saving data to the config file.