Saving changed settings to User Config File
During this working arc I've created a custom User Settings class, as well as a custom Data Interaction Helper in purpose to change the Settings regardless of it's type. For now I've implemented the saving only for Game Difficulty param and tested that it works perfectly. Also I've implemented the saving of the Changed Setting to Config File, so now even if we restart the game, our setting would still be the same.
Next task is to work on the visual part of the Settings and implement Highlighting for it.