Highlighting of a selected/hovered Option
During this arc I've implemented the Highlight of the ListEntry after we Hover our mouse over it, or returning to Default style if we stop hovering over it. Also implemented the highlight of the selected ListEntry. Now if we click on one of our ListEntry it will be highlighted and the one that was highlighted before would return to default style. Also fixed the problem with the selection: Now if we change the property inside Rotator this ListEntry would be selected and also if we just click on the Rotator it would be selected as well.
Next task is to start the implementation of the Details View Widget.