Details View Widget
During this arc I've implemented the Details View Widget. Now if go to the Options and select ListEntry, on the right side of it we will see the description of current ListEntry. This description is customizable and it's  content can be changed by developer. Also implemented the feature with showing Details View Widget of the item, that user currently hovering over and if we do not hover some ListEntry we show Details View of currently selected ListEntry.
Next task is to implement the functionality of the Reset Button.